requirements.txt
hazelcast-python-client>=5.6.0,<6.0.0
